ClientChainData (a series of device info from client). Feel free to get xuid or more device-local information.
